Fixed a bug that would result in blurry canvas.

+++ web/src/main/kotlin/world/phantasmal/web/core/rendering/InputManager.kt
@@ -4,7 +4,7 @@ package world.phantasmal.web.core.rendering
  * Manages user input such as pointer and keyboard events.
  */
 interface InputManager {
-    fun setSize(width: Double, height: Double)
+    fun setSize(width: Int, height: Int)
 
     fun resetCamera()
